Small reorganisation in Domain (details below)
Removed ProgramFileManager.py
Moved ProgramFileStructure.py to Domain
Moved utility classes such as Helpers.py, SDFHandler.py and XSDCreator.py to a new Domain.util folder
