+++ Domain/ProgramFileStructure.py
@@ -1,4 +1,7 @@
 from pathlib import Path
+import sys
+
+from Domain.logger.OTLLogger import OTLLogger
 
 
 class ProgramFileStructure:
@@ -31,3 +34,14 @@ class ProgramFileStructure:
             # cls.download_fresh_otlmow_model(model_dir_path)
             # cls.get_otlmow_model_version(model_dir_path)
         return model_dir_path
+
+    @classmethod
+    def get_dynamic_library_path(cls, library_name: str) -> Path:
+        dynamic_library_path = Path(library_name)
+        if hasattr(sys, '_MEIPASS'):  # when in .exe file
+            dynamic_library_path = Path(sys._MEIPASS, library_name)
+        elif not dynamic_library_path.exists():
+            dynamic_library_path = Path('data', library_name)
+        OTLLogger.logger.debug(f"Supplied dynamic library path: {dynamic_library_path}")
+
+        return dynamic_library_path
