+++ Domain/util/Helpers.py
