+++ GUI/dialog_windows/ExportProjectWindow.py
@@ -3,7 +3,6 @@ from pathlib import Path
 from PyQt6.QtWidgets import QFileDialog
 
 from Domain.project.Project import Project
-from Domain.project.ProgramFileManager import ProgramFileManager
 
 
 class ExportProjectWindow:
