+++ UnitTests/general_fixtures/DomainFixtures.py
@@ -5,14 +5,12 @@ from pathlib import Path
 from unittest.mock import Mock, AsyncMock
 
 from _pytest.fixtures import fixture
-from typing_extensions import NamedTuple
 
 from Domain import global_vars
 from Domain.logger.OTLLogger import OTLLogger
-from Domain.project.ProgramFileStructure import ProgramFileStructure
+from Domain.ProgramFileStructure import ProgramFileStructure
 from Domain.step_domain.InsertDataDomain import InsertDataDomain
 from Domain.project.Project import Project
-from Domain.project.ProgramFileManager import ProgramFileManager
 from Domain.step_domain.RelationChangeDomain import RelationChangeDomain
 
 OTLLogger.logger = Mock()
