+++ csv_to_sqlite.py
@@ -1,11 +1,9 @@
-import csv
 import sqlite3
 from pathlib import Path
 
 import otlmow_converter.OtlmowConverter
-from urllib3.util.proxy import connection_requires_http_tunnel
 
-from Domain.Helpers import Helpers
+from Domain.util.Helpers import Helpers
 from Domain.logger.OTLLogger import OTLLogger
 from Exceptions.NotASqlliteFileError import NotASqlliteFileError
